feat(ref-klv-04): mirror ST 0903 u64 widening into JVM + Python bindings
JNI: add checked_u64 (rejects negative i64, passes any non-negative as u64
since Java has no unsigned primitive), swap checked_u32 -> checked_u64 for
targetId + centroidPixel + bboxTopLeft/BottomRight + centroidPixRow/Col on
decode; encode uses target_id as i64 (i64::from(u64) doesn't compile).
Python: extract::<u32> -> extract::<u64> for the same 5+1 fields in klv.rs;
errors.rs VTargetPackEmpty/DuplicateTargetId cast u64 as u32 for the
informational .tag attribute (full info in message). Java record was already
long/Long; .pyi already int — no surface change, no C ABI change. Large-value
round-trip tests on both: V6 fields (target_id + Tags 1/2/3) tested above
u32::MAX; V4 fields (Tags 19/20) capped at u32::MAX per spec §10.2.2.20/.21.

+++ bindings/python/src/klv.rs
@@ -715,7 +715,7 @@ fn py_to_vtarget_pack(p: &Bound<'_, PyAny>) -> PyResult<RustVTargetPack> {
     let py = p.py();
 
     // BER-OID `target_id` (mandatory, no Tag).
-    r.target_id = p.getattr(intern!(py, "target_id"))?.extract::<u32>()?;
+    r.target_id = p.getattr(intern!(py, "target_id"))?.extract::<u64>()?;
 
     macro_rules! op {
         ($field:ident, $ty:ty) => {
@@ -738,9 +738,9 @@ fn py_to_vtarget_pack(p: &Bound<'_, PyAny>) -> PyResult<RustVTargetPack> {
         };
     }
 
-    op!(centroid_pixel, u32);
-    op!(bbox_top_left_pixel, u32);
-    op!(bbox_bottom_right_pixel, u32);
+    op!(centroid_pixel, u64);
+    op!(bbox_top_left_pixel, u64);
+    op!(bbox_bottom_right_pixel, u64);
     op!(priority, u8);
     op!(confidence_level, u8);
     op!(history, u16);
@@ -772,8 +772,8 @@ fn py_to_vtarget_pack(p: &Bound<'_, PyAny>) -> PyResult<RustVTargetPack> {
     op!(bbox_bottom_right_lon_offset, f64);
     ob!(target_location);
     ob!(geospatial_contour_series);
-    op!(centroid_pix_row, u32);
-    op!(centroid_pix_col, u32);
+    op!(centroid_pix_row, u64);
+    op!(centroid_pix_col, u64);
     op!(algorithm_id, u32);
     op!(detection_status, u8);
     ob!(vmask);
